Fix extension loaded multiple times in Linux
The code checks whether or not the library has already been loaded by
looking at the path() of every loaded dlops object.

For Linux ".so" is appended before loading the library, but this is not
appended when checking whether it's loaded. This causes the library to
get added to the dlops list multiple times.

This patch fixes the issue by introducing a new variable that has the
required platform dependent suffix and uses that for the comparisons and
loading.

The `name` variable stays unchanged so that the logging messages uses
the extension name as it appears in sqf.